Move users management to its own app.
