+++ templates/users/change_password.html
@@ -1,4 +1,4 @@
-{% extends 'user/_user.html' %}
+{% extends 'users/_user.html' %}
 {% load form_helpers %}
 {% block title %}Change Password{% endblock %}
 {% block user_content %}
@@ -22,7 +22,7 @@
             </div>
             <div class="text-right">
               <button type="submit" name="_update" class="btn btn-primary">Update</button>
-              <a href="{% url 'user_profile' %}" class="btn btn-secondary">Cancel</a>
+              <a href="{% url 'users:profile' %}" class="btn btn-secondary">Cancel</a>
             </div>
           </form>
 {% endblock %}
