+++ templates/users/profile.html
@@ -1,4 +1,4 @@
-{% extends 'user/_user.html' %}
+{% extends 'users/_user.html' %}
 {% block title %}User Profile{% endblock %}
 {% block user_content %}
           <small class="text-muted">User login</small>
